backend/requirements.txt
Django==4.2.10
django-cors-headers==3.8.0
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.2.0
django-shortuuidfield==0.1.3
django-two-factor-auth==1.15.5
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
channels==3.0.1
python-logstash==0.4.8
uvicorn==0.27.1
python-jose==3.3.0
psycopg2==2.8.6
postgres==4.0
websockets==12.0
whitenoise==6.6.0